- ActorStore - using an optional MovieStore so we don't need to initialize it until its being used - ActorsStore - added CreateActor( ) , deleteAllSavedActors() -  ActorsStore - updated fetch functions - CastStore - can fetch for specific movie cast, will save first 25 cast members to disk - MovieCard - added movieRating property to update rating entity - not working - MovieRatingStore - updated fetching functions

+++ PopcornSwirl/Home.swift
@@ -13,6 +13,8 @@ import Alamofire
 struct Home: View {
     
     var movieRatings = MovieRatingStore()
+//    var movieRatings : MovieRatingStore?
+    
     @ObservedObject private var genreStore = GenreStore()
 
     @ObservedObject var movieStore = MovieStore()
@@ -120,16 +122,31 @@ struct Home: View {
     } // Nav
         
     .onAppear() {
+        
+        // TMDB
         movieStore.getGenres()
         movieStore.fetchPopularMovies()
         movieStore.fetchUpcomingMovies()
         
-        movieRatings.fetchAllRatings()
-//        movieRatings.fetchRatingsForMovie(id: 25)
+        // CoreData
+        genreStore.fetchAllGenres()
+        
+        
+        
+        
+//        movieRatings.deleteAllMovieRatings()
+            movieRatings.fetchAllRatings()
+        guard let s = movieRatings.fetchRatingsForMovie(id: 1) else { return }
+        
+        
+        
+        let actorsStore = ActorsStore()
+        actorsStore.deleteAllSavedActors() 
+        
         let castStore = CastStore()
         castStore.deleteAll()
         
-        genreStore.fetchAllGenres()
+        
         
     }
